% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaguar_math.R
\name{addme}
\alias{addme}
\title{Add together two numbers.}
\usage{
addme(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
addme(1, 1)
addme(10, 1)
}

