% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{generate_code}
\alias{generate_code}
\title{Generate code from data frames
This is not part of the data pipeline, but used to more quickly create
code that is used on the service-side of the application.  If columns
change in the data frames or new data frames are added this can be
run to update the code.}
\usage{
generate_code()
}
\description{
This function will place several files in the BASE_PATH.
}

